+++ tests/floris_interface_integration_test.py
@@ -16,6 +16,20 @@ def test_read_yaml():
     fi = FlorisInterface(configuration=YAML_INPUT)
     assert isinstance(fi, FlorisInterface)
 
+def test_assign_yaw_angles():
+
+    fi = FlorisInterface(configuration=YAML_INPUT)
+    fi.set(layout_x=[0, 0], layout_y=[0, 1000])
+
+    # Test setting yaw angles via a list
+    fi.set(yaw_angles=[20.0, 30.0])
+
+    # Test passing in integers as yaw angles
+    fi.set(yaw_angles=[20, 30])
+
+    # Test setting yaw angles via a numpy array
+    fi.set(yaw_angles=np.array([20.0, 30.0]))
+
 def test_set_run():
     """
     These tests are designed to test the set / run sequence to ensure that inputs are
